# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dienstplan-App API für Persönliche Assistenz
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: Alle Verträge auflisten
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Liste der Verträge
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Vertrag anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Vertrag abrufen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vertrag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Nicht gefunden
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Vertrag aktualisieren
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Vertrag löschen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Gelöscht
...
  /shift-models:
    get:
      operationId: listShiftModels
      tags: [shift-models]
      summary: Schichtmodelle auflisten
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Liste der Schichtmodelle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftModel"
    post:
      operationId: createShiftModel
      tags: [shift-models]
      summary: Schichtmodell anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftModelInput"
      responses:
        "201":
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftModel"
  /shift-models/{id}:
    patch:
      operationId: updateShiftModel
      tags: [shift-models]
      summary: Schichtmodell aktualisieren
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftModelUpdate"
      responses:
        "200":
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftModel"
        "404":
          description: Nicht gefunden
    delete:
      operationId: deleteShiftModel
      tags: [shift-models]
      summary: Schichtmodell löschen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Gelöscht
        "404":
          description: Nicht gefunden
...
  /contracts/{id}/vacation-balance:
    get:
      operationId: getVacationBalance
      tags: [contracts]
      summary: Urlaubskontigent abrufen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Urlaubskontigent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacationBalance"
        "404":
          description: Nicht gefunden
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Anmelden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Erfolgreich angemeldet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Falsche Zugangsdaten
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Aktuell eingeloggter Benutzer
      responses:
        "200":
          description: Eingeloggter Benutzer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Nicht angemeldet
  /auth/set-password:
    post:
      operationId: setPassword
      tags: [auth]
      summary: Passwort über Einladungstoken setzen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordInput"
      responses:
        "200":
          description: Passwort gesetzt und angemeldet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Ungültiger oder abgelaufener Token
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, accountType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, assistant]
        accountType:
          type: string
          enum: [privat, dienstleister]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        socialSecurityNumber:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        taxClass:
          type: ["string", "null"]
        healthInsurance:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Contract:
      type: object
      required: [id, userId, weeklyHours, vacationDays, vacationDaysUsed, startDate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        weeklyHours:
          type: number
        vacationDays:
          type: integer
        vacationDaysUsed:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    ContractInput:
      type: object
      required: [userId, weeklyHours, vacationDays, startDate]
      properties:
        userId:
          type: integer
        weeklyHours:
          type: number
          minimum: 0
        vacationDays:
          type: integer
          minimum: 0
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
    ContractUpdate:
      type: object
      properties:
        weeklyHours:
          type: number
          minimum: 0
        vacationDays:
          type: integer
          minimum: 0
        vacationDaysUsed:
          type: integer
          minimum: 0
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    ShiftModel:
      type: object
      required: [id, name, valuationPercent, color, sortOrder, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        valuationPercent:
          type: integer
        color:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ShiftModelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        valuationPercent:
          type: integer
          minimum: 0
        color:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
    ShiftModelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        valuationPercent:
          type: integer
          minimum: 0
        color:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
    VacationBalance:
      type: object
      required: [contractId, userId, vacationDays, vacationDaysUsed, vacationDaysRemaining]
      properties:
        contractId:
          type: integer
        userId:
          type: integer
        vacationDays:
          type: integer
        vacationDaysUsed:
          type: integer
        vacationDaysRemaining:
          type: integer
    AuthUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, assistant]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8