# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dienstplan-App API für Persönliche Assistenz
 * OpenAPI spec version: 0.1.0
 */
import type { UserAccountType } from './userAccountType';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  accountType: UserAccountType;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  socialSecurityNumber?: string | null;
  /** @nullable */
  taxId?: string | null;
  /** @nullable */
  taxClass?: string | null;
  /** @nullable */
  healthInsurance?: string | null;
  /** @nullable */
  iban?: string | null;
  isActive?: boolean;
  createdAt: Date;
}